+++ arthive-backend/app/graphql/types/review_type.rb
@@ -2,7 +2,7 @@ module Types
     class ReviewType < Types::BaseObject
         field :id, ID, null: false
         field :content, String, null: true
-        field :rating, Int, null: false
+        field :rating, Float, null: false
         field :if_favorite, Boolean, null: false
         field :if_finished, Boolean, null: false
         field :created_at, GraphQL::Types::ISO8601DateTime, null: false
